feat: add traefik configuration for reverse proxy and SSL feat: add compose.yml for docker services configuration remove: delete old docker-compose.yml file
The changes were made to introduce Traefik as a reverse proxy and SSL manager for the services. The new compose.yml file is added to manage the configuration of all docker services in a more organized way. The old docker-compose.yml file is removed as it is no longer needed.